Refactor to use only actions in the router and move logic for defaulting to the latest thread to the openThread action
